+++ voreen/modules/openmp/src/voreenblasmp.cpp
@@ -75,7 +75,7 @@ void VoreenBlasMP::sSpMVEll(const EllpackMatrix<float>& mat, const float* vec, f
         result[row] = 0.f;
         for (size_t colIndex=0; colIndex < numColsPerRow; ++colIndex) {
             size_t col = mat.getColumn(row, colIndex);
-            if(col != -1) {
+            if(col != static_cast<size_t>(-1)) {
                 result[row] += mat.getValueByIndex(row, colIndex) * vec[col];
             }
         }
@@ -94,7 +94,7 @@ void VoreenBlasMP::hSpMVEll(const EllpackMatrix<int16_t>& mat, const float* vec,
         result[row] = 0.f;
         for (size_t colIndex=0; colIndex < numColsPerRow; ++colIndex) {
             size_t col = mat.getColumn(row, colIndex);
-            if(col != -1) {
+            if(col != static_cast<size_t>(-1)) {
                 result[row] += mat.getValueByIndex(row, colIndex) * vec[col];
             }
         }
@@ -113,7 +113,7 @@ float VoreenBlasMP::sSpInnerProductEll(const EllpackMatrix<float>& mat, const fl
         float dot = 0.f;
         for (size_t colIndex=0; colIndex < numColsPerRow; ++colIndex) {
             size_t col = mat.getColumn(row, colIndex);
-            if(col != -1) {
+            if(col != static_cast<size_t>(-1)) {
                 dot += mat.getValueByIndex(row, colIndex) * vecy[col];
             }
         }
